Fixed cross references.
Z<> tags must have an empty line after them, lest any subsequent X<>
tags make a paragraph. This is a flaw in the PseudoPOD formatter, but
it's easier to fix by convention for now than it is to change in the
formatter. (In other words, it's kind of a limitation of POD.)